+++ pages/review/index.vue
@@ -1,0 +1,7 @@
+<template></template>
+<script setup lang="ts">
+const router = useRouter();
+
+// redirect to the home page
+router.push("/");
+</script>
